Be stricter about what we mean by 'marked' X.
If we say that a PMD entry is marked dirty, it is clear - the dirty flag is set
in this entry. However if we conflate things by saying the PTE page it's
pointing at is 'marked dirty', it's confusing - the page itself isn't marked,
the entry is. Better to just say the PTE page is dirty and if mentioning
'marking' refer to the PMD entry.