Update functions to correctly refer to gigantic/huge page arrangements.
If a PMD entry is marked with the PSE flag, then its physical address refers to
the physical page rather than a PTE page. Equally, if a PUD entry is marked PSE,
it refers to a gigantic physical page.

This is necessary because we need more bits for the offset into the physical
page.