Fix confusion about page table allocation.
It turns out that the first and last page table entries observed are purely due
to malloc() book keeping - mmap() does not create this, it is not a kernel thing
at all. This makes sense as - why use the memory if we don't need it yet? The
VMA will protect the memory range.